codec derive - crate code split into multiple files
